% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chooseEnrichmentRows}
\alias{chooseEnrichmentRows}
\title{Choose Enrichment Rows}
\usage{
chooseEnrichmentRows(dataTable, maxDefault = 30)
}
\arguments{
\item{dataTable}{A data.table object holding results from enrichment, such as from
\code{enrichmentAnalysisREACTOMEWS} or \code{enrichmentAnalysisDAVID}.}

\item{maxDefault}{The maximum number of rows to use if there are many significant results.}
}
\value{
the shortened version of the provided dataTable.
}
\description{
Processes an enrichment analysis data.table (which contains a pvalue and qvalue rows)
to produce a selection of the rows, considering that a very high number of rows would
be problematic for most people in an HTML report.
}

