% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/proteinSetReport.R
\name{convertBaderLabGMTForPiano}
\alias{convertBaderLabGMTForPiano}
\title{Convert GMT Bader Lab For Piano}
\usage{
convertBaderLabGMTForPiano(pathToGMTFile)
}
\arguments{
\item{pathToGMTFile}{The path to the .gmt file obtained from Bader lab. The file needs
to be ended with .gmt extension.}
}
\value{
the path to the converted file (same file, but .gmt extension changed to .piano.gmt)
}
\description{
Adapts the a Bader Lab (http://download.baderlab.org/EM_Genesets/) gene set file (gmt)
so that the gene set can be used by Piano.
}

