% Generated by roxygen2 (4.0.2): do not edit by hand
\name{enrichmentAnalysisDAVID}
\alias{enrichmentAnalysisDAVID}
\title{Enrichment analysis DAVID}
\usage{
enrichmentAnalysisDAVID(davidEmail, proteinList, listName,
  prot2GeneName = list())
}
\arguments{
\item{davidEmail}{The email registered at DAVID to be able to use their web service.}

\item{proteinList}{The list of proteins (UniProt Accessions) that should be analyzed.}

\item{listName}{A label for the list.}

\item{prot2GeneName}{An optional dictionary (in a data.table) with prot and GeneName columns,
used to translate the uniprot accessions to Gene names.}
}
\value{
the resulting enrichment analysis (functional chart) from DAVID, as a data.table.
}
\description{
\code{enrichmentAnalysisDAVID} uses the \code{RDAVIDWebService} package to execute an
enrichment analysis over http with the DAVID server.
}

