% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/proteinSetReport.R
\name{getENTREZGIFromUNIPROTIDs}
\alias{getENTREZGIFromUNIPROTIDs}
\title{Get ENTREZ GI From ENSEMBL IDs}
\usage{
getENTREZGIFromUNIPROTIDs(martUniprot, martEnsembl, uniprotIDs)
}
\arguments{
\item{martUniprot}{The uniprot biomart instance to use}

\item{martEnsembl}{The ensembl biomart instance to use}

\item{ensemblIDs}{A vector of UniProt IDs to get ENTREZ GIs}
}
\value{
A data.table view of the biomart result, with \code{accession} and \code{entrezgene}.
}
\description{
This function uses a biomart connection to UniProt to fetch proteins to ensembl gene ids, and a
ENSEMBL biomart to fetch ENSEMBL gene ids to ENTREZ GI.
}

