% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getUniprotProteinNames}
\alias{getUniprotProteinNames}
\title{Get UniProt Protein Names}
\usage{
getUniprotProteinNames(ensemblIDs, species = "mmusculus_gene_ensembl")
}
\arguments{
\item{ensemblIDs}{The identifiers to search names for. These should belong to the
data set specifief by \code{species}}

\item{species}{The name of the ENSEMBL Biomart data set to use}
}
\value{
A data.table with rows \code{ensembl_gene_id} and \code{uniprot_genename}.
}
\description{
Retrieves the UniProt Gene name for the provided ENSEMBL identifiers. The ENSEMBL
Biomart data set is choosen through the species parameter. Multiple names are collapsed
into a comma separated list, so that ENSEMBL gene ids are unique.
}

