% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/proteinSetReport.R
\name{makeInterproLink}
\alias{makeInterproLink}
\title{Make InterPro link}
\usage{
makeInterproLink(interproEntry, dict = NA, type = "html")
}
\arguments{
\item{interproEntry}{The accession of the UniProt entry.}

\item{dict}{The dictionary (data frame with entry and name columns) to
translate the entry into something different for the URL label. Defaults to \code{NA}.}

\item{type}{Either "html" or "markdown", defining the type of link generated. Defaults to "html".}
}
\value{
The link for the InterPro domain entry.
}
\description{
Produces a link to InterPro for the specified entry.
}
\seealso{
\code{\link{makeLinkForDB}}
}

