% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/proteinSetReport.R
\name{makeInterProtDomainArchLink}
\alias{makeInterProtDomainArchLink}
\title{Make InterPro protein architecture (UniProt) link}
\usage{
makeInterProtDomainArchLink(uniprotEntry, type = "html")
}
\arguments{
\item{uniprotEntry}{The accession of the UniProt entry.}

\item{type}{Either "html" or "markdown", defining the type of link generated. Defaults to "html".}
}
\value{
The link to the InterPro page for the UniProt entry.
}
\description{
Produces a link to InterPro for a UniProt specified entry.
}
\seealso{
\code{\link{makeLinkForDB}}
}

