% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/proteinSetReport.R
\name{makeLinkForDB}
\alias{makeLinkForDB}
\title{Make link for DB}
\usage{
makeLinkForDB(entry, urlPrefix, urlPostfix = "", dict, type)
}
\arguments{
\item{entry}{Identifier of the database where we are linking to}

\item{urlPrefix}{The prefix to be used before the entry for constructing the URL.}

\item{urlPostfix}{The postfix to be used after the entry for constructing the URL. Defaults to an empty string.}

\item{dict}{An optional data.frame which has two columns \code{entry} and \code{name}, and acts as a dictionary to
translate the identifier in entry to something else (such as a Gene name) to be used as label of the link.}

\item{type}{Either "html" or "markdown", defining the type of link generated.}
}
\value{
the string with the link.
}
\description{
Lower level function of the link making series.
}

