% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/proteinSetReport.R
\name{makeMolIntOntologyLink}
\alias{makeMolIntOntologyLink}
\title{Make Molecular Interaction Ontology Link}
\usage{
makeMolIntOntologyLink(line, type = "html")
}
\arguments{
\item{line}{A line containing the accession of the MI Ontology followed by the description.}

\item{type}{Either "html" or "markdown", defining the type of link generated. Defaults to "html".}

\item{dict}{The dictionary (data frame with entry and name columns) to
translate the entry into something different for the URL label. Defaults to \code{NA}.}
}
\value{
The link for the MI.
}
\description{
\code{makeMolIntOntologyLink} Produces a link to the Ontology Look-up Service (OLS) from the EBI,
for the specified entry of the MI Ontology.
}
\details{
This function should probably be improved to receive a dictionary instead of the description in the
input.
}
\seealso{
\code{\link{makeLinkForDB}}
}

