% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/proteinSetReport.R
\name{mergeGenesToGSARes}
\alias{mergeGenesToGSARes}
\title{Merge Genes to GSA Res}
\usage{
mergeGenesToGSARes(gsaRes, padjCutoff = 0.5, species = NA)
}
\arguments{
\item{gsaRes}{A piano GSA result object.}

\item{padjCutoff}{A cut-off for rows to show in the final table; defaults to \code{0.5}.}

\item{species}{The ensembl biomart data set name for retrieving names for the genes. This will only work
if the GSA piano method was executed with a data set where the identifier for genes are ENSEMBL IDs.}
}
\description{
Piano GSA result summary table doesn't include the detail of which genes in the query are
part of each enriched category/group. This method joins the summary to the genes that belong
to each category. It can cut the list if a cut-off for the adjusted p-value is given.
}

