% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/proteinSetReport.R
\name{obtainDataFromMart}
\alias{obtainDataFromMart}
\title{Obtain data from mart}
\usage{
obtainDataFromMart(mart, filters, attributes, values, key)
}
\arguments{
\item{mart}{the biomart object previously initialized.}

\item{filters}{the set of filters to be used on the biomart call.}

\item{attributes}{the set of attributes to be used on the biomart call.}

\item{values}{the query values to be used, in equivalent dimension as the filters.}

\item{key}{the column to be used as key in the data.table, this should be part of the attributes}
}
\value{
The biomart result as a data.table indexed by key.
}
\description{
\code{obtainDataFromMart} encapsulates a biomart call to produce a data.table object with the result.
}

