% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/proteinSetReport.R
\name{obtainUniprotLinks}
\alias{obtainUniprotLinks}
\title{obtain UniProt Links}
\usage{
obtainUniprotLinks(mart, field, values)
}
\arguments{
\item{mart}{The biomart connection to use.}

\item{field}{The field to query for.}

\item{values}{A vector with the identifiers to query for.}
}
\value{
A data.table containing the query to uniprot accession (and the database they belong to)
}
\description{
Produces a list of uniprot identifiers for the given search field and values.
The search field can be ENSEMBL_Transcripts, ENSEMBL_Genes, and then the values
would be vectors containing either transcript identifiers or genes.
}

