% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/proteinSetReport.R
\name{parseDavidKEGGPathwayTerm}
\alias{parseDavidKEGGPathwayTerm}
\title{Parse DAVID KEGG Pathway Term}
\usage{
parseDavidKEGGPathwayTerm(term)
}
\arguments{
\item{term}{An individual term from the DAVID enrichment analysis. This is not a vectorized function.}
}
\value{
A named list with species, pathID and pathName. These are the required fields for the pathview function,
besides the list of proteins to colour.
}
\description{
Parses a "term" from the result of a DAVID enrichment analysis, as produced by function \code{enrichmentAnalysisDAVID},
to extract KEGG Pathways enriched to be feed to the pathview function to colour KEGG Pathway diagrams.
}

