% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/proteinSetReport.R
\name{uniprotEntrySource}
\alias{uniprotEntrySource}
\title{Uniprot entry source}
\usage{
uniprotEntrySource(unimart, prots)
}
\arguments{
\item{unimart}{The UniProt biomart to use for this operation}

\item{prots}{A vector/list of proteins to query for}
}
\value{
A data.table containing the identifier - source database mapping (Accession,Status)
}
\description{
Given a set of UniProt identifiers, it return the same list indicating whether the entries
belong to Swissprot or TrEMBL.
}

