% Generated by roxygen2 (4.0.2): do not edit by hand
\name{enrichmentAnalysisREACTOME}
\alias{enrichmentAnalysisREACTOME}
\title{Enrichment Analysis REACTOME}
\usage{
enrichmentAnalysisREACTOME(giList, gi2GeneName = list())
}
\arguments{
\item{giList}{A vector of entrez gene identifiers to run the analysis for}

\item{gi2GeneName}{An optional dictionary (as a data.table) with columns gi \code{gi} to \code{GeneName},
which maps entrez gene ids to gene names.}
}
\value{
The Reactome enrichment analysis for the vector of gis given, as a data.table
}
\description{
Runs enrichment analysis using the ReactomePA.
}

