% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/proteinSetReport.R
\name{enrichmentAnalysisREACTOMEWS}
\alias{enrichmentAnalysisREACTOMEWS}
\title{Enrichment Analysis REACTOME WS}
\usage{
enrichmentAnalysisREACTOMEWS(proteins)
}
\arguments{
\item{proteins}{A vector of UniProt identifiers to run the analysis for}
}
\value{
The Reactome enrichment analysis for the vector of proteins given, as a data.table
}
\description{
Runs enrichment analysis using the REACTOME REST web service.
}

