% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/proteinSetReport.R
\name{getENTREZGIFromENSEMBLIDs}
\alias{getENTREZGIFromENSEMBLIDs}
\title{Get ENTREZ GI From ENSEMBL IDs}
\usage{
getENTREZGIFromENSEMBLIDs(mart, ensemblIDs)
}
\arguments{
\item{mart}{The biomart (ENSEMBL) instance to use}

\item{ensemblIDs}{A vector of ENSEMBL IDs to get ENTREZ GIs}
}
\value{
A data.table view of the biomart result, with \code{ensembl_gene_id} as key and \code{entrezgene}.
}
\description{
This function uses a biomart connection to ENSEMBL biomart to fetch ENSEMBL gene ids to
ENTREZ GI.
}

